Move registry manager
