+++ src/windows-emulator/registry/registry_manager.cpp
