+++ src/windows-emulator/registry/registry_manager.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "std_include.hpp"
+#include "../std_include.hpp"
 #include <serialization_helper.hpp>
 
 class hive_parser;
